added db interface
