+++ app/src/main/java/com/shipsmart/data/dbLogic/AndroidUser.kt
@@ -9,7 +9,7 @@ import kotlinx.serialization.Serializable
 @Entity (
     tableName = "users",
     indices = [Index("id")]
-    )
+)
 data class AndroidUser(
     @PrimaryKey(autoGenerate = false) var id: Int? = null,
     @ColumnInfo(name = "Email") var userEmail: String? = null,
@@ -18,12 +18,4 @@ data class AndroidUser(
 //    @ColumnInfo(name = "Name") var userName: String? = null,
 //    @ColumnInfo(name = "Email") var userEmail: String? = null,
 //    @ColumnInfo(name = "Password") var userPassword: String? = null
-)
-
-@Serializable
-data class SupabaseUser(
-    var id: Int? = null,
-    var name: String? = null,
-    var email: String,
-    var password: String
 )
