+++ app/src/main/java/com/shipsmart/data/dbLogic/SupabaseUser.kt
@@ -1,0 +1,11 @@
+package com.shipsmart.data.dbLogic
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class SupabaseUser(
+    var id: Int? = null,
+    var name: String? = null,
+    var email: String,
+    var password: String
+)
