+++ app/src/main/java/com/shipsmart/domain/dbDAO.kt
@@ -1,0 +1,8 @@
+package com.shipsmart.domain
+
+import com.shipsmart.data.dbLogic.SupabaseUser
+
+interface dbDAO {
+    suspend fun getUser(email: String) : SupabaseUser?
+    suspend fun addUser(user: SupabaseUser)
+}
